Add files via upload
Having a list of n numbers with digits in range [0, S], where n <= 100, switch all list positions in O(n) time. If the input number contains a digit greater or equal than S, you will delete the digit from the number, for example with S=6, 61 becomes 1, and 6 will be deleted from the array. The result should be printed in console/terminal.